+++ src/ploomber/tasks/notebook.py
@@ -823,6 +823,37 @@ class ScriptRunner(NotebookMixin, Task):
     it also works by injecting a cell into the source code. Source can be
     a ``.py`` script or an ``.ipynb`` notebook. **Does not support magics.**
 
+    Parameters
+    ----------
+    source: str or pathlib.Path
+        Script source, if str, the content is interpreted as the actual
+        script, if pathlib.Path, the content of the file is loaded. When
+        loading from a str, ext_in must be passed
+    product: ploomber.File
+        The output file
+    dag: ploomber.DAG
+        A DAG to add this task to
+    name: str, optional
+        A str to indentify this task. Should not already exist in the dag
+    params: dict, optional
+        Script parameters. This are passed as the "parameters" argument
+        to the papermill.execute_notebook function, by default, "product"
+        and "upstream" are included
+    ext_in: str, optional
+        Source extension. Required if loading from a str. If source is a
+        ``pathlib.Path``, the extension from the file is used.
+    static_analysis : ('disabled', 'regular', 'strict'), default='regular'
+        Check for various errors in the script. In 'regular' mode, it aborts
+        execution if the notebook has syntax issues, or similar problems that
+        would cause the code to break if executed. In 'strict' mode, it
+        performs the same checks but raises an issue before starting execution
+        of any task, furthermore, it verifies that the parameters cell and
+        the params passed to the notebook match, thus, making the script
+        behave like a function with a signature.
+    local_execution : bool, optional
+        Change working directory to be the parent of the script source.
+        Defaults to False.
+
     Examples
     --------
 
@@ -863,12 +894,12 @@ class ScriptRunner(NotebookMixin, Task):
         local_execution=False,
     ):
         self.ext_in = ext_in
-        self.local_execution = local_execution
 
         kwargs = dict(hot_reload=dag._params.hot_reload)
         self._source = ScriptRunner._init_source(
             source, kwargs, ext_in, static_analysis, False, False
         )
+        self.local_execution = local_execution
         super().__init__(product, dag, name, params)
 
     @staticmethod
@@ -897,11 +928,6 @@ class ScriptRunner(NotebookMixin, Task):
         if self.static_analysis == "regular":
             self.source._check_notebook(raise_=True, check_signature=False)
 
-        if self.local_execution:
-            cwd = str(self.source.loc.parent)
-        else:
-            cwd = None
-
         fd, tmp = tempfile.mkstemp(".py")
         os.close(fd)
 
@@ -912,6 +938,14 @@ class ScriptRunner(NotebookMixin, Task):
                 if c["cell_type"] == "code"
             ]
         )
+
+        cwd = str(self.source.loc.parent.resolve())
+        orig_env = os.environ.copy()
+
+        if "PYTHONPATH" not in orig_env:
+            orig_env["PYTHONPATH"] = cwd
+        else:
+            orig_env["PYTHONPATH"] += os.pathsep + cwd
 
         tmp = Path(tmp)
         tmp.write_text(code)
@@ -924,15 +958,15 @@ class ScriptRunner(NotebookMixin, Task):
             raise ValueError("ScriptRunner only works with Python and R scripts")
 
         try:
-            _run_script_in_subprocess(interpreter, tmp, cwd)
+            _run_script_in_subprocess(interpreter, tmp, cwd, orig_env)
         except Exception as e:
             raise TaskBuildError("Error when executing task" f" {self.name!r}.") from e
         finally:
             tmp.unlink()
 
 
-def _run_script_in_subprocess(interpreter, path, cwd):
-    res = subprocess.run([interpreter, str(path)], cwd=cwd, stderr=PIPE)
+def _run_script_in_subprocess(interpreter, path, cwd, env):
+    res = subprocess.run([interpreter, str(path)], cwd=cwd, env=env, stderr=PIPE)
 
     if res.returncode:
         stderr = res.stderr.decode()