+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * [Fix] Clearer error messages when generating pipeline plots
 * [Fix] Fix error when choosing which backend plot to use (when user didn't expicitly requested one)
 
+* [Fix] Fixed local execution error when using Scriptrunner
+
 ## 0.22.2 (2023-01-30)
 
 * [Fix] Fixes error that printed a git error message when failing to retrieve current git hash ([#1067](https://github.com/ploomber/ploomber/issues/1067))